task模块 接口编写完成

+++ codes/back-end/app/v1/api/task.py
@@ -2,67 +2,63 @@
 @Auth ： youngZ
 @File ：task.py
 """
-from typing import Any
 from core import deps
 from fastapi import APIRouter, Depends
 from models import User, Task
-from scheams import TaskPydantic, TaskInPydantic, ResponseSuccess, ResponseFailed
+from scheams import TaskPydantic, TaskInPydantic, TaskWithPublisherPydantic
+from scheams import ResponseSuccess, ResponseFailed
 
 task = APIRouter(tags=["任务模块"], prefix="/task")
 
 
-@task.get("/list", summary="任务列表")  # 路径host:port/tms/task/list
-async def get_task_list(limit: int = 10, page: int = 1):
+@task.get("/list", summary="查询所有任务")  # 路径host:port/tms/task/list
+async def get_task_list(limit: int = 10, page: int = 1, user_obj: User = Depends(deps.get_current_user)):
     skip = (page - 1) * limit
-    try:
-        data = {
-            "total": await Task.all().count(),
-            "tasks": await TaskPydantic.from_queryset(Task.all().select_related("publisher").offset(skip).limit(limit).order_by('id'))
-        }
-    except Exception as error:
-        return ResponseFailed(data=error, message="查询失败")
-    else:
-        return ResponseSuccess(data=data, message="查询成功")
+    data = {
+        "total": await Task.all().count(),
+        "tasks": await TaskWithPublisherPydantic.from_queryset(
+            Task.all().prefetch_related("publisher").offset(skip).limit(limit).order_by('id'))
+    }
+
+    return ResponseSuccess(message="查询成功", data=data)
 
 
-@task.get("/info/{id}", summary="任务信息")  # 路径host:port/tms/task/info/1
-async def get_task_by_id(id: int):
-    try:
-        data = await Task.filter(id=id)
-    except Exception as error:
-        return ResponseFailed(data=error, message="查询失败")
-    else:
-        return ResponseSuccess(data=data, message="查询成功")
+@task.get("/info/{id}", summary="查询任务详细信息")  # 路径host:port/tms/task/info/1
+async def get_task_by_id(id: int, user_obj: User = Depends(deps.get_current_user)):
+    exist: bool = await Task.filter(id=id).exists()
+    if not exist:
+        return ResponseFailed(message="任务不存在", data=None)
+
+    query_task = await TaskWithPublisherPydantic.from_tortoise_orm(await Task.get(id=id).prefetch_related("publisher"))
+    return ResponseSuccess(message="查询成功", data=query_task)
 
 
-@task.post("/create", summary="新增任务", )  # 路径host:port/tms/task/create
-async def create_task(task_form: TaskInPydantic, user: User = Depends(deps.get_current_user)):
-    try:
-        data = await TaskInPydantic.from_tortoise_orm(
-            await Task.create(**task_form.dict(), publisher=user)
-        )
-    except Exception as error:
-        return ResponseFailed(data=str(error), message="新增失败")
-    else:
-        return ResponseSuccess(data=data, message="新增成功")
+@task.post("/create", summary="新增一条任务")  # 路径host:port/tms/task/create
+async def create_task(task_form: TaskInPydantic, user_obj: User = Depends(deps.get_current_user)):
+    created_task = await TaskInPydantic.from_tortoise_orm(await Task.create(**task_form.dict(), publisher=user_obj))
+
+    return ResponseSuccess(message="新增成功", data=created_task)
 
 
 @task.put("/update/{id}", summary="修改任务")  # 路径host:port/tms/task/update/1
-async def update_task_by_id(id: int, task_form: TaskInPydantic):
-    try:
-        data = await Task.filter(id=id).update(**task_form.dict())
+async def update_task_by_id(id: int, task_form: TaskInPydantic, user_obj: User = Depends(deps.get_current_user)):
+    exist: bool = await Task.filter(id=id).exists()
+    if not exist:
+        return ResponseFailed(message="任务不存在", data=None)
 
-    except Exception as error:
-        return ResponseFailed(data=error, message="修改失败")
-    else:
-        return ResponseSuccess(data=data, message="修改成功")
+    old_task = await TaskPydantic.from_tortoise_orm(await Task.get(id=id))
+    await Task.filter(id=id).update(**task_form.dict())  # 执行修改操作
+    new_task = await TaskPydantic.from_tortoise_orm(await Task.get(id=id))
+
+    return ResponseSuccess(message="修改成功", data={"old task": old_task, "new task": new_task})
 
 
 @task.delete("/delete/{id}", summary="删除任务")  # 路径host:port/tms/task/delete/1
-async def delete_task_by_id(id: int):
-    try:
-        data = await Task.filter(id=id).delete()
-    except Exception as error:
-        return ResponseFailed(data=error, message="删除失败")
-    else:
-        return ResponseSuccess(data=data, message="删除成功")
+async def delete_task_by_id(id: int, user_obj: User = Depends(deps.get_current_user)):
+    exist: bool = await Task.filter(id=id).exists()
+    if not exist:
+        return ResponseFailed(message="删除失败，任务不存在", data=None)
+
+    delete_count = await Task.filter(id=id).delete()
+
+    return ResponseSuccess(message="删除成功", data={"delete count": delete_count})
